Fix crash when terminating pre-launched Lacros, if already terminated
Other calls to Terminate inside BrowserManager always check that
lacros_process_ is valid before terminating it. Not doing so may
result in timing issues in which Ash (and pre-launched Lacros) is
terminated before we get into ResumeLaunch, causing a crash.

Bug: 1446157
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4569945
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1150164}

+++ chrome/browser/ash/crosapi/browser_manager.cc
@@ -1549,7 +1549,9 @@ void BrowserManager::ResumeLaunch() {
     LOG(WARNING) << "Lacros is not enabled for the current user. Terminating "
                     "pre-launched instance";
     unload_requested_ = true;
-    lacros_process_.Terminate(/*exit_code=*/0, /*wait=*/false);
+    if (lacros_process_.IsValid()) {
+      lacros_process_.Terminate(/*exit_code=*/0, /*wait=*/false);
+    }
     return;
   }
